# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Train Booking API",
    "version": "1.0.0",
    "description": "API for managing train schedules, routes, and ticket bookings"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current Server"
    }
  ],
  "paths": {
    "/api/bookings": {
      "get": {
        "summary": "Retrieve all bookings",
        "tags": [
          "Bookings"
        ],
        "responses": {
          "200": {
            "description": "A list of all customer bookings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "bookingCode": {
                        "type": "string"
                      },
                      "scheduleId": {
                        "type": "string"
                      },
                      "routeId": {
                        "type": "string"
                      },
                      "ticketClass": {
                        "type": "string"
                      },
                      "selectedDay": {
                        "type": "string"
                      },
                      "totalAmount": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/trips": {
      "get": {
        "summary": "Get all trips",
        "tags": [
          "Trips"
        ],
        "responses": {
          "200": {
            "description": "A list of trips.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trip"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unable to retrieve trips.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/trips/{id}": {
      "get": {
        "summary": "Get a trip by ID",
        "tags": [
          "Trips"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique trip identifier."
          }
        ],
        "responses": {
          "200": {
            "description": "The requested trip.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "404": {
            "description": "Trip not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unable to retrieve trip.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/schedules": {
      "get": {
        "summary": "Get all schedules",
        "tags": [
          "Schedules"
        ],
        "responses": {
          "200": {
            "description": "A list of schedules.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a schedule",
        "tags": [
          "Schedules"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Schedule"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Schedule created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/schedules/{id}": {
      "get": {
        "summary": "Get a schedule by ID",
        "tags": [
          "Schedules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Numeric schedule identifier.",
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "The requested schedule.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Schedule"
                }
              }
            }
          },
          "404": {
            "description": "Schedule not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a schedule",
        "tags": [
          "Schedules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Numeric schedule identifier.",
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Schedule"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Schedule updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "404": {
            "description": "Schedule not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a schedule",
        "tags": [
          "Schedules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Numeric schedule identifier.",
            "schema": {
              "type": "integer"
            },
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Schedule deleted successfully."
          },
          "404": {
            "description": "Schedule not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/schedules/trip/{tripId}": {
      "get": {
        "summary": "Get schedules for a trip",
        "tags": [
          "Schedules"
        ],
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "description": "Trip identifier.",
            "schema": {
              "type": "string"
            },
            "example": "alpine-panorama"
          }
        ],
        "responses": {
          "200": {
            "description": "Schedules belonging to the trip.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/schedules/trip/{tripId}/active": {
      "get": {
        "summary": "Get active schedules for a trip",
        "tags": [
          "Schedules"
        ],
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "description": "Trip identifier.",
            "schema": {
              "type": "string"
            },
            "example": "alpine-panorama"
          }
        ],
        "responses": {
          "200": {
            "description": "Active schedules belonging to the trip.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Schedule"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Trip": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "startStation": {
            "type": "string"
          },
          "endStation": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "distance": {
            "type": "number"
          },
          "highlights": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "bestSeason": {
            "type": "string"
          },
          "operatingMonths": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "imageUrl": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Schedule not found"
          }
        }
      },
      "Schedule": {
        "type": "object",
        "required": [
          "id",
          "tripId",
          "departureTime",
          "arrivalTime",
          "daysOfWeek"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique numeric identifier for the schedule.",
            "example": 1
          },
          "tripId": {
            "type": "string",
            "description": "Identifier of the trip served by this schedule.",
            "example": "alpine-panorama"
          },
          "departureTime": {
            "type": "string",
            "description": "Departure time in HH:mm format.",
            "example": "08:30"
          },
          "arrivalTime": {
            "type": "string",
            "description": "Arrival time in HH:mm format.",
            "example": "13:00"
          },
          "daysOfWeek": {
            "type": "array",
            "description": "Days on which the schedule operates.",
            "items": {
              "type": "string"
            },
            "example": [
              "monday",
              "tuesday",
              "wednesday",
              "thursday",
              "friday"
            ]
          },
          "status": {
            "type": "boolean",
            "description": "Whether the schedule is active.",
            "default": true,
            "example": true
          }
        }
      },
      "ScheduleResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Schedule created successfully"
          },
          "data": {
            "$ref": "#/components/schemas/Schedule"
          }
        }
      }
    }
  },
  "tags": []
}